Implemented the new Auth flow Improved the ProfileListing update event
