+++ .env.production
@@ -8,3 +8,4 @@ REACT_APP_MARKETPLACE_NAME=The Luupe - QA
 REACT_APP_MARKETPLACE_ROOT_URL=https://market.qa.theluupe.com
 REACT_APP_SHARETRIBE_SDK_CLIENT_ID=198e4d3e-b64d-4e19-a503-8309affbbd5c
 REACT_APP_STRIPE_PUBLISHABLE_KEY=pk_test_51PGPLOFibVw716Tl9JCOjMHX5Py936BJbfkpiu2mjdcnMgtySmcEwVJMa2wF58GhKyV0bRVf6MKxIlFytF7Lrsmq00X3iAdO2i
+REACT_APP_MARKETPLACE_BRANDING_BUCKET=branding-dev.theluupe.com
