+++ server/api/auth/loginWithIdp.js
@@ -38,7 +38,7 @@ module.exports = (err, user, req, res, idpClientId, idpId) => {
           maxAge: 15 * 60 * 1000, // 15 minutes
         }
       )
-      .redirect(`${rootUrl}/login#`);
+      .redirect(`${rootUrl}/signup#`);
   }
 
   if (!user) {
@@ -61,7 +61,7 @@ module.exports = (err, user, req, res, idpClientId, idpId) => {
           maxAge: 15 * 60 * 1000, // 15 minutes
         }
       )
-      .redirect(`${rootUrl}/login#`);
+      .redirect(`${rootUrl}/signup#`);
   }
 
   const { from, defaultReturn, defaultConfirm, userType } = user;
