+++ src/containers/AuthenticationPage/ConfirmSignupForm/ConfirmSignupForm.module.css
@@ -24,11 +24,11 @@
 }
 
 .firstNameRoot {
-  width: calc(34% - 9px);
+  width: calc(40% - 9px);
 }
 
 .lastNameRoot {
-  width: calc(66% - 9px);
+  width: calc(60% - 9px);
 }
 
 .password {
