--- src/containers/AuthenticationPage/LoginForm/LoginForm.example.js
@@ -1,13 +1,0 @@
-/* eslint-disable no-console */
-import LoginForm from './LoginForm';
-
-export const Empty = {
-  component: LoginForm,
-  props: {
-    formId: 'LoginFormExample',
-    onSubmit(values) {
-      console.log('log in with form values:', values);
-    },
-  },
-  group: 'page:AuthenticationPage',
-};
