--- src/containers/AuthenticationPage/SignupForm/SignupForm.example.js
@@ -1,113 +1,0 @@
-import React from 'react';
-import { fakeIntl } from '../../../util/testData';
-import TermsAndConditions from '../TermsAndConditions/TermsAndConditions';
-import SignupForm from './SignupForm';
-
-const userTypes = [
-  {
-    userType: 'a',
-    label: 'Seller',
-  },
-  {
-    userType: 'b',
-    label: 'Buyer',
-  },
-  {
-    userType: 'c',
-    label: 'Guest',
-  },
-  {
-    userType: 'd',
-    label: 'Host',
-  },
-];
-
-const userFields = [
-  {
-    key: 'enumField1',
-    scope: 'public',
-    schemaType: 'enum',
-    enumOptions: [
-      { option: 'o1', label: 'l1' },
-      { option: 'o2', label: 'l2' },
-      { option: 'o3', label: 'l3' },
-    ],
-    saveConfig: {
-      label: 'Enum Field 1',
-      displayInSignUp: true,
-      isRequired: false,
-    },
-    userTypeConfig: {
-      limitToUserTypeIds: false,
-    },
-  },
-  {
-    key: 'enumField2',
-    scope: 'public',
-    schemaType: 'enum',
-    enumOptions: [
-      { option: 'o1', label: 'l1' },
-      { option: 'o2', label: 'l2' },
-      { option: 'o3', label: 'l3' },
-    ],
-    saveConfig: {
-      label: 'Enum Field 2',
-      displayInSignUp: true,
-      isRequired: false,
-    },
-    userTypeConfig: {
-      limitToUserTypeIds: true,
-      userTypeIds: ['c', 'd'],
-    },
-  },
-  {
-    key: 'textField',
-    scope: 'private',
-    schemaType: 'text',
-    saveConfig: {
-      label: 'Text Field',
-      displayInSignUp: true,
-      isRequired: true,
-    },
-    userTypeConfig: {
-      limitToUserTypeIds: false,
-    },
-  },
-  {
-    key: 'booleanField',
-    scope: 'protected',
-    schemaType: 'boolean',
-    saveConfig: {
-      label: 'Boolean Field',
-      displayInSignUp: true,
-      isRequired: false,
-    },
-    userTypeConfig: {
-      limitToUserTypeIds: false,
-    },
-  },
-];
-
-export const Empty = {
-  component: SignupForm,
-  props: {
-    formId: 'SignupFormExample',
-    userTypes,
-    userFields,
-    onSubmit(values) {
-      console.log('sign up with form values:', values);
-    },
-    onOpenTermsOfService() {
-      console.log('open terms of service');
-    },
-    termsAndConditions: (
-      <TermsAndConditions
-        onOpenTermsOfService={() => setTosModalOpen(true)}
-        onOpenPrivacyPolicy={() => setPrivacyModalOpen(true)}
-        formId="SignupFormExample"
-        intl={fakeIntl}
-      />
-    ),
-  },
-  group: 'page:AuthenticationPage',
-};
