--- src/containers/AuthenticationPage/socialLoginLogos.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-// This is Auth0's logo, you are not allowed to change its color
-export const Auth0Logo = (
-  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 64 64">
-    <path
-      d="M49.012 51.774L42.514 32l17.008-12.22h-21.02L32.005 0h21.032l6.506 19.78c3.767 11.468-.118 24.52-10.53 31.993zm-34.023 0L31.998 64l17.015-12.226-17.008-12.22zm-10.516-32c-3.976 12.1.64 24.917 10.5 32.007v-.007L21.482 32 4.474 19.774l21.025.007L31.998 0H10.972z"
-      fill="#eb5424"
-    />
-  </svg>
-);
